fix: failing builds on cluster
This was caused by two problems:

The node no longer has the correct QEMU
configuration after a reboot. Adding a step
that sets it first ensures it never is incorrect.

The Dockerfile had a bug that only occured with
Arm. Revert the 2 separate RUN commands to a known
to work RUN command from the original source.